Move CStadium traits
