+++ include/football/CStadium.h
@@ -1,31 +1,11 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CStadium.h"
 
 #include "types/CStadium.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CStadium;
-}
-
-template <> struct json_traits<football::CStadium>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Stadium";
-	//! JSON key for the \copybrief football::CStadium::mName
-	static inline constexpr std::string_view NAME_KEY = "Name";
-	//! JSON key for the \copybrief football::CStadium::mCapacity
-	static inline constexpr std::string_view CAPACITY_KEY = "Capacity";
-	//! JSON key for the \copybrief football::CStadium::mAmbientFactor
-	static inline constexpr std::string_view AMBIENT_FACTOR_KEY = "Ambient factor";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -82,6 +62,4 @@ private:
 	ambient_factor mAmbientFactor;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
